# docs/guides/2026-03-10-支撑系统接口文档-openapi3.1.yaml
...
openapi: 3.1.0
info:
  title: EleTender 支撑系统接口文档
  description: 基于运行中 ele-tender-support 服务自动生成，并补充关键枚举与中文说明。
  contact:
    name: EleTender
  version: 1.0.0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
...
  /api/auth/change-password:
    post:
      tags:
      - 认证管理
      summary: 修改当前用户密码
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/auth/info:
    get:
      tags:
      - 认证管理
      summary: 获取当前用户信息（别名）
      operationId: getUserInfoAlias
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultUserInfo'
  /api/auth/login:
    post:
      tags:
      - 认证管理
      summary: 用户登录
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultUserLoginResponse'
  /api/auth/logout:
    post:
      tags:
      - 认证管理
      summary: 用户登出
      operationId: logout
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultVoid'
  /api/auth/menus:
    get:
      tags:
      - 认证管理
      summary: 获取当前用户菜单
      operationId: getUserMenus
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultListSysMenu'
  /api/auth/userinfo:
    get:
      tags:
      - 认证管理
      summary: 获取当前用户信息
      operationId: getUserInfo
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultUserInfo'
...
  /api/external/token:
    post:
      tags:
      - 外部系统认证
      summary: 外部系统用户换取Token
      operationId: getToken
      parameters:
      - name: X-App-Key
        in: header
        description: 应用Key
        required: true
        schema:
          type: string
      - name: X-Timestamp
        in: header
        description: 时间戳
        required: true
        schema:
          type: integer
          format: int64
      - name: X-Signature
        in: header
        description: 签名
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalTokenRequest'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultExternalTokenResponse'
  /api/external/userinfo:
    get:
      tags:
      - 外部系统认证
      summary: 获取当前外部用户信息
      operationId: getExternalUserInfo
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultExternalUserInfoResponse'
  /api/external/verify:
    get:
      tags:
      - 外部系统认证
      summary: 验证外部系统签名
      operationId: verify
      parameters:
      - name: X-App-Key
        in: header
        description: 应用Key
        required: true
        schema:
          type: string
      - name: X-Timestamp
        in: header
        description: 时间戳
        required: true
        schema:
          type: integer
          format: int64
      - name: X-Signature
        in: header
        description: 签名
        required: true
        schema:
          type: string
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultBoolean'
  /api/menus:
    get:
      tags:
      - 菜单管理
      summary: 获取菜单列表
      operationId: getMenuList
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultListSysMenu'
    post:
      tags:
      - 菜单管理
      summary: 创建菜单
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SysMenu'
        required: true
      responses:
        '500':
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultObject'
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultSysMenu'
...
components:
  schemas:
    ResultObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    ResultVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    SysMenu:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createTime:
          type: string
          format: date-time
        createId:
          type: integer
          format: int64
        createName:
          type: string
        modifyTime:
          type: string
          format: date-time
        modifyId:
          type: integer
          format: int64
        modifyName:
          type: string
        ver:
          type: integer
          format: int32
        isDelete:
          type: integer
          format: int32
        parentId:
          type: integer
          description: 父菜单ID，0表示根菜单
          format: int64
        menuName:
          type: string
          description: 菜单名称
        menuCode:
          type: string
          description: 菜单编码
        menuType:
          type: integer
          description: '菜单类型: 1-菜单, 2-按钮'
          format: int32
          enum:
          - 1
          - 2
          x-enumDescriptions:
          - 菜单
          - 按钮
        menuUrl:
          type: string
          description: 菜单URL
        permission:
          type: string
          description: 权限标识
        icon:
          type: string
          description: 图标
        sortOrder:
          type: integer
          description: 排序号
          format: int32
        status:
          type: integer
          description: '状态: 0-禁用, 1-启用'
          format: int32
          enum:
          - 0
          - 1
          x-enumDescriptions:
          - 禁用
          - 启用
        children:
          type: array
          description: 子菜单
          items:
            $ref: '#/components/schemas/SysMenu'
      description: 系统菜单
...
    ResultSysMenu:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/SysMenu'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    ExternalTokenRequest:
      type: object
      properties:
        userName:
          type: string
        userId:
          type: string
        enterpriseName:
          type: string
        enterpriseId:
          type: string
        enterpriseCode:
          type: string
    ExternalTokenResponse:
      type: object
      properties:
        token:
          type: string
          description: Token
        expireIn:
          type: integer
          description: 过期时间（秒）
          format: int64
        tokenType:
          type: string
          description: Token类型
      description: 外部系统Token响应
    ResultExternalTokenResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ExternalTokenResponse'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    UserLoginRequest:
      required:
      - password
      - username
      type: object
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          description: 密码
      description: 用户登录请求
    ResultUserLoginResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserLoginResponse'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    UserInfo:
      type: object
      properties:
        userId:
          type: integer
          description: 用户ID
          format: int64
        username:
          type: string
          description: 用户名
        realName:
          type: string
          description: 真实姓名
        email:
          type: string
          description: 邮箱
        phone:
          type: string
          description: 手机号
        roles:
          type: array
          description: 角色列表
          items:
            type: string
            description: 角色列表
      description: 用户信息
    UserLoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Token
        expireIn:
          type: integer
          description: 过期时间（秒）
          format: int64
        userInfo:
          $ref: '#/components/schemas/UserInfo'
        permissions:
          type: array
          description: 权限列表
          items:
            type: string
            description: 权限列表
      description: 用户登录响应
...
    ResultListSysMenu:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/SysMenu'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    ResultBoolean:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: boolean
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
    ExternalUserInfoResponse:
      type: object
      properties:
        appKey:
          type: string
          description: 应用Key
        userId:
          type: string
          description: 外部用户ID
        userName:
          type: string
          description: 外部用户名称
        enterpriseId:
          type: string
          description: 企业ID
        enterpriseName:
          type: string
          description: 企业名称
        enterpriseCode:
          type: string
          description: 企业社会统一信用代码
      description: 外部用户信息响应
    ResultExternalUserInfoResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ExternalUserInfoResponse'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...
    ResultUserInfo:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserInfo'
        timestamp:
          type: integer
          format: int64
        success:
          type: boolean
...